% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altText.R
\name{altText}
\alias{altText}
\title{Returns alternate text for Parameter name.}
\usage{
altText(lookUpParameter)
}
\arguments{
\item{lookUpParameter}{parameter for which alternate text is required}
}
\value{
Alternate text as a string
}
\description{
Returns alternate text for Parameter name.
}
\examples{
altText("TDS")

}
