% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censored.R
\name{censored}
\alias{censored}
\title{Takes Character and returns if censored left or right}
\usage{
censored(Value)
}
\arguments{
\item{Value}{character string with numbers}
}
\value{
logical
}
\description{
Takes Character and returns if censored left or right
}
\examples{
censored("<100")
censored(">100")




}
