% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvalues.R
\docType{data}
\name{cvalues}
\alias{cvalues}
\title{Control values for closed systems}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18 rows and 4 columns.
}
\usage{
data(cvalues)
}
\description{
Control values for closed systems
}
\keyword{datasets}
