% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateExcel.R
\name{dateExcel}
\alias{dateExcel}
\title{Takes Excel date (numbered from 1899-12-30) returns ISO date}
\usage{
dateExcel(excelDate)
}
\arguments{
\item{excelDate}{number}
}
\value{
Date
}
\description{
Takes Excel date (numbered from 1899-12-30) returns ISO date
}
\examples{
dateExcel(42705)




}
