% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateUnix.R
\name{dateUnix}
\alias{dateUnix}
\title{Converts a UNIX timestamp to an ISO date}
\usage{
dateUnix(value)
}
\arguments{
\item{value}{number}
}
\value{
Date
}
\description{
Converts a UNIX timestamp to an ISO date
}
\examples{
dateUnix(1669766400)




}
