% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotlyParam.R
\name{ggplotlyParam}
\alias{ggplotlyParam}
\title{Plot a timeline of closed system parameter option to make a plotly graph}
\usage{
ggplotlyParam(
  df,
  site = "",
  plot_system = "LPHW",
  plot_parameter = "pH",
  gtitle = "",
  data_from = 2010,
  show_controlvalues = F,
  linear_trend = F,
  loess_trend = T,
  graph_facet = "portrait",
  graph_wrap = T,
  show_caption = F,
  caption_text = "",
  date_scale = 3,
  date_period = "months",
  date_form = "\%b \%y",
  note_label = F,
  gg_plotly = F,
  inhibitor = "mixed"
)
}
\arguments{
\item{df}{Data frame with Date, SampleID, value and Parameter column.}

\item{site}{String site name blank by default}

\item{plot_system}{String. System name blank by default.}

\item{plot_parameter}{String. parameter name}

\item{gtitle}{String, graph title}

\item{data_from}{Integer start Year of data}

\item{show_controlvalues}{Logical. Display control limits. FALSE by default.}

\item{linear_trend}{Logical. Display linear trend}

\item{loess_trend}{Logical. Display loess trend}

\item{graph_facet}{String "portrait" or "landscape", portrait by default.}

\item{graph_wrap}{Logical. If TRUE graphs are faceted. TRUE by default.}

\item{show_caption}{Logical FALSE by default.}

\item{caption_text}{String caption text "" by default}

\item{date_scale}{Integer number of months/weeks between Date axis values}

\item{date_period}{String. e.g. "months" or "weeks"}

\item{date_form}{String. Date format string}

\item{note_label}{Logical FALSE by default.}

\item{gg_plotly}{Logical. Make plotly graph FALSE by default.}

\item{inhibitor}{String. Inhibitor strategy; "nitrite", "mixed"  by default.}
}
\value{
ggplot or ggplotly graph
}
\description{
Plot a timeline of closed system parameter option to make a plotly graph
and option to make a control value graph
}
\examples{
Date <- c("2022-10-3", "2022-11-7", "2022-12-5", "2023-01-3", "2023-02-7", "2023-03-5")
Date <- lubridate::ymd(Date)
Year <- lubridate::year(Date)
Parameter <- c("pH", "pH", "pH", "pH", "pH", "pH")
value <- c(7.5, 8.8, 9.0, 5.0, 6.5, 7.8)
system <- c("LPHW", "LPHW", "LPHW", "LPHW", "LPHW", "LPHW")
SampleID <- c("LPHW 1", "LPHW 1", "LPHW 1", "LPHW 1", "LPHW 1", "LPHW 1")
system_df <- tibble::tibble(Date, Year, Parameter, value, system, SampleID)
ggplotlyParam(system_df, site = "No 10", plot_parameter = "pH")

ggplotlyParam(system_df, site = "No 10", plot_parameter = "pH", gg_plotly = T)

Date <- c("2022-08-3", "2022-10-7", "2022-12-5", "2023-02-3", "2023-04-7")
Date <- lubridate::ymd(Date)
Year <- lubridate::year(Date)
Parameter <- c("Glycol", "Glycol", "Glycol", "Glycol", "Glycol")
value <- c(15, 20, 30, 29, 31)
system <- c("CHW", "CHW", "CHW", "CHW", "CHW")
SampleID <- c("CHW 1", "CHW 1", "CHW 1", "CHW 1", "CHW 1")
system_df <- tibble::tibble(Date, Year, Parameter, value, system, SampleID)
ggplotlyParam(system_df,
              site = "No 10",
              plot_system = "CHW",
              plot_parameter = "Glycol",
              show_controlvalues = T)

}
