% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotlyParam3.R
\name{ggplotlyParam3}
\alias{ggplotlyParam3}
\title{Plot a timeline of a location parameter option to make a plotly graph}
\usage{
ggplotlyParam3(
  df,
  site = "",
  PlotLocation = "Polishing Pond",
  PlotParm = "",
  gtitle = "",
  axsisTxt = "",
  figTxt = "",
  ylim1 = 0,
  ylim2 = 600,
  show_controlvalues = F,
  linear_trend = F,
  loess_trend = T,
  loess_se = T,
  graph_facet = "portrait",
  graph_wrap = T,
  show_caption = F,
  caption_text = "",
  date_scale = 4,
  date_period = "months",
  date_form = "\%b \%y",
  note_label = F,
  gg_plotly = F,
  inhibitor = "mixed",
  Soluble = F,
  Caption = "",
  txtSize = 16
)
}
\arguments{
\item{df}{Data frame with Date, SampleID, value and Parameter column.}

\item{site}{String site name blank by default}

\item{PlotLocation}{String. Location name blank by default.}

\item{PlotParm}{String. parameter name}

\item{gtitle}{String, graph title}

\item{axsisTxt}{String, axsis title}

\item{figTxt}{String, fig title}

\item{ylim1}{numeric lower limit y axsis}

\item{ylim2}{numeric upper limit y axsis}

\item{show_controlvalues}{Logical. Display control limits. FALSE by default.}

\item{linear_trend}{Logical. Display linear trend}

\item{loess_trend}{Logical. Display loess trend}

\item{graph_facet}{String "portrait" or "landscape", portrait by default.}

\item{graph_wrap}{Logical. If TRUE graphs are faceted. TRUE by default.}

\item{show_caption}{Logical FALSE by default.}

\item{caption_text}{String caption text "" by default}

\item{date_scale}{Integer number of months/weeks between Date axis values}

\item{date_period}{String. e.g. "months" or "weeks"}

\item{date_form}{String. Date format string}

\item{note_label}{Logical FALSE by default.}

\item{gg_plotly}{Logical. Make plotly graph FALSE by default.}

\item{inhibitor}{String. Inhibitor strategy; "nitrite", "mixed"  by default.}

\item{Soluble}{Logical. If TRUE soluble values shown as a different shape.}

\item{Caption}{String, text for the graph caption}

\item{txtSize}{numeric, theme base_font size}
}
\value{
ggplot or ggplotly graph
}
\description{
Plot a timeline of a location parameter option to make a plotly graph
and option to make a control value graph
}
\examples{
Date <- c("2022-10-3", "2022-11-7", "2022-12-5", "2023-01-3", "2023-02-7", "2023-03-5")
Date <- lubridate::ymd(Date)
Year <- lubridate::year(Date)
Parameter <- c("pH", "pH", "pH", "pH", "pH", "pH")
value <- c(7.5, 8.8, 9.0, 5.0, 6.5, 7.8)
Location <- c("LPHW", "LPHW", "LPHW", "LPHW", "LPHW", "LPHW")
SampleID <- c("LPHW 1", "LPHW 1", "LPHW 1", "LPHW 1", "LPHW 1", "LPHW 1")
system_df <- tibble::tibble(Date, Year, Parameter, value, Location, SampleID)
ggplotlyParam3(system_df, site = "No 10", PlotParm = "pH")

ggplotlyParam3(system_df, site = "No 10", PlotParm = "pH", gg_plotly = T)

Date <- c("2022-08-3", "2022-10-7", "2022-12-5", "2023-02-3", "2023-04-7")
Date <- lubridate::ymd(Date)
Year <- lubridate::year(Date)
Parameter <- c("Glycol", "Glycol", "Glycol", "Glycol", "Glycol")
value <- c(15, 20, 30, 29, 31)
Location <- c("CHW", "CHW", "CHW", "CHW", "CHW")
SampleID <- c("CHW 1", "CHW 1", "CHW 1", "CHW 1", "CHW 1")
system_df <- tibble::tibble(Date, Year, Parameter, value, Location, SampleID)
ggplotlyParam3(system_df,
              site = "No 10",
              PlotLocation = "CHW",
              PlotParm = "Glycol",
              show_controlvalues = T)

Date <- c("2022-08-3", "2022-10-7", "2022-12-5", "2023-02-3", "2023-04-7")
Date <- lubridate::ymd(Date)
Year <- lubridate::year(Date)
Parameter <- c("Pseudomonas Species", "Pseudomonas Species", "Pseudomonas Species", "Pseudomonas Species", "Pseudomonas Species")
value <- c(100, 2000, 50000, 1000000, 300000)
Location <- c("CHW", "CHW", "CHW", "CHW", "CHW")
SampleID <- c("CHW 1", "CHW 1", "CHW 1", "CHW 1", "CHW 1")
system_df <- tibble::tibble(Date, Year, Parameter, value, Location, SampleID)
ggplotlyParam3(system_df,
              site = "No 10",
              PlotLocation = "CHW",
              PlotParm = "Pseudomonas Species",
              show_controlvalues = F,
              loess_trend = F)

}
