% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphParameter.R
\name{graphParameter}
\alias{graphParameter}
\title{Takes a Latis name and makes a consistent parameter name}
\usage{
graphParameter(df, g_param = "pH", g_facet = T)
}
\arguments{
\item{df}{data frame with Date, SampleID, Value and Parameter column}
}
\value{
ggplot graph
}
\description{
Latis lab names only
}
\examples{
Date <- c("2022-10-3", "2022-11-7", "2022-12-5", "2022-10-3", "2022-11-7", "2022-12-5")
Date <- lubridate::ymd(Date)
Parameter <- c("pH", "pH", "pH", "pH", "pH", "pH")
value <- c(7.5, 8.8, 9.0, 5.0, 6.5, 7.8)
SampleID <- c("CHW 1", "CHW 1", "CHW 1", "CHW 2", "CHW 2", "CHW 2")
system_df <- tibble(Date, Parameter, value, SampleID)
graphParameter(system_df, g_param = "pH", g_facet = T)

}
