% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leftCen.R
\name{leftCen}
\alias{leftCen}
\title{Takes Character and returns if left censored}
\usage{
leftCen(Value)
}
\arguments{
\item{Value}{character string with numbers}
}
\value{
logical
}
\description{
Takes Character and returns if left censored
}
\examples{
leftCen("<100")




}
