% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelFactor.R
\name{levelFactor}
\alias{levelFactor}
\title{Takes column and shows the levels when made a factor}
\usage{
levelFactor(Value)
}
\arguments{
\item{Value}{character string}
}
\value{
character vector
}
\description{
Takes column and shows the levels when made a factor
}
\examples{
exampleVector <- c("aa", "bb", "cc", "aa", "dd", "cc", "cc")
levelFactor(exampleVector)



}
