% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNum.R
\name{MakeNum}
\alias{MakeNum}
\title{Takes Character and returns Number}
\usage{
MakeNum(Value)
}
\arguments{
\item{Value}{character with numbers}
}
\value{
numeric value
}
\description{
Takes Character and returns Number
}
\examples{
MakeNum("<100")




}
