% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massBalance.R
\name{massBalance}
\alias{massBalance}
\title{Returns concentration of pollutant downstream}
\usage{
massBalance(Fr, Cr, fd, cd)
}
\arguments{
\item{Fr}{- river flow upstream of the discharge}

\item{Cr}{- concentration of the pollutant in the river upstream of the discharge}

\item{fd}{- flow of the discharge}

\item{cd}{- concentration of the pollutant in the discharge}
}
\value{
Concentration of the pollutant downstream of the discharge
}
\description{
Returns concentration of pollutant downstream
}
\examples{
massBalance(25, 0.06, 2, 5.5)

}
