% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameParameter.R
\name{nameParameter}
\alias{nameParameter}
\title{Takes a Latis name and makes a consistent parameter name}
\usage{
nameParameter(df)
}
\arguments{
\item{df}{data frame with LatisName column}
}
\value{
Parameter
}
\description{
Latis lab names only
}
\examples{
latis_df <- tibble(LatisName = c("Appearance", "ChloridG01", "Cu36", "Fe36" , TVC22/72hr))
nameParameter(latis_df)

}
