% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize a numerical vector between 0-1}
\usage{
normalize(x)
}
\arguments{
\item{Value}{numerical vector}
}
\value{
numerical vector with values between 0 and 1.
}
\description{
Normalize a numerical vector between 0-1
}
\examples{
normalize(c(1, 2, 3, 4, 5))




}
