% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramColour.R
\name{paramColour}
\alias{paramColour}
\title{Returns colour of Parameter to show on the graph.}
\usage{
paramColour(lookUpParameter)
}
\arguments{
\item{lookUpParameter}{parameter for which colour is required}
}
\value{
Colour as a string
}
\description{
Returns colour of Parameter to show on the graph.
}
\examples{
paramColour("Conductivity")

}
