% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramList.R
\name{paramList}
\alias{paramList}
\title{Returns a vector of parameters that can be graphed with graphParameter}
\usage{
paramList()
}
\value{
vector of parameters
}
\description{
Returns a vector of parameters that can be graphed with graphParameter
function. Used to check if the parameter can be plotted.
}
\examples{
paramList()

}
