% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramUnits.R
\name{paramUnits}
\alias{paramUnits}
\title{Parameter name and returns the units of measurement.}
\usage{
paramUnits(lookUpParameter)
}
\arguments{
\item{lookUpParameter}{parameter for which units are required}
}
\value{
Units of measuremenmt as a string
}
\description{
Parameter name and returns the units of measurement.
}
\examples{
paramUnits("Conductivity")

}
