% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerCarbon.R
\name{powerCarbon}
\alias{powerCarbon}
\title{Returns CO2 emissions from power (kW)}
\usage{
powerCarbon(kW, factor = 0.527)
}
\arguments{
\item{kW}{eletrical energy (kWh)}

\item{factor}{conversion factor (kWh to kg) 2021 default}
}
\value{
carbon dioxide (kg)
}
\description{
based on 527 tonnes of carbon dioxide per gigawatt hour (GWh) overall in 2021
Bulb has a figure 0f 0.193 kg per kWh
}
\examples{
powerCarbon(40)
powerCarbon(35, factor = 0.193)

}
