% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rightCen.R
\name{rightCen}
\alias{rightCen}
\title{Takes Character and returns if right censored}
\usage{
rightCen(Value)
}
\arguments{
\item{Value}{character string with numbers}
}
\value{
logical
}
\description{
Takes Character and returns if right censored
}
\examples{
rightCen(">100000")




}
