% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterHeatingkWh.R
\name{waterHeatingkWh}
\alias{waterHeatingkWh}
\title{Returns kWh needed to change the temperature of water}
\usage{
waterHeatingkWh(m, T_i, T_f, eff = 0.88)
}
\arguments{
\item{m}{- the mass of water - tonnes}

\item{T_i}{- the initial temperature}

\item{T_f}{- the final temperature}

\item{Eff}{- Efficiency of the boiler}
}
\value{
kWh for heater water from an initial to final temperature
}
\description{
Returns kWh needed to change the temperature of water
}
\examples{
waterHeatingkWh(68, 15, 60, eff = 0.88)
waterHeatingkWh(3536, 55, 56, eff = 0.88)

}
