% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorYrMn.R
\name{factorYrMn}
\alias{factorYrMn}
\title{Returns factor levels for Month (abbrev)Year in order (4 or 2 digit years)}
\usage{
factorYrMn(fromYear, toYear, yr_type = 2)
}
\arguments{
\item{fromYear}{integer Year}

\item{toYear}{integer Year}

\item{yr_type}{2 or 4 digit years}
}
\value{
character vector
}
\description{
Returns factor levels for Month (abbrev)Year in order (4 or 2 digit years)
}
\examples{
factorYrMn(2000, 2023)


}
