% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNum.R
\name{makeNum}
\alias{makeNum}
\title{Takes Character and returns Number}
\usage{
makeNum(Value)
}
\arguments{
\item{Value}{character string with numbers}
}
\value{
numeric value
}
\description{
Takes Character and returns Number
}
\examples{
makeNum("<100")
makeNum(" 5,811.0")




}
