% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDate.R
\name{metaDate}
\alias{metaDate}
\title{Add date information}
\usage{
metaDate(system_df)
}
\arguments{
\item{system_df}{data frame with Date column}
}
\value{
data.frame
}
\description{
From the Date column creates useful date information (Month, MonthNr, MonthFull, Year, Week,
YearDay, MonthDay, WeekDay, WeekDayFull, Season); as extra columns in the data frame.
NOTE: a POSIXct vector will be converted to a Date vector.
}
\details{
Season:
Meteorological spring March 1 to May 31.
Meteorological summer June 1 to August 31.
Meteorological autumn September 1 to November 30.
Meteorological winter December 1 to February 28.
}
\examples{
sys_data_frame <- tibble(Date = Sys.Date())
metaDate(sys_data_frame)

}
