% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerCarbon.R
\name{powerCarbon}
\alias{powerCarbon}
\title{Returns CO2 emissions from power (kW)}
\usage{
powerCarbon(kW, factor = 0.193)
}
\arguments{
\item{kW}{eletrical energy (kWh)}

\item{factor}{conversion factor (kWh to kg) 2021 default}
}
\value{
carbon dioxide (kg)
}
\description{
In 2022 (Bulb), these figures were 0.193 kg of CO2e per kWh of electricity
Department for Environment, Food & Rural Affairs (DEFRA) is 0.233 kgCO2e per
kWh of electricity consumed 2020
}
\examples{
powerCarbon(40)
powerCarbon(35, factor = 0.233)

}
